GW.AscendPlan/src/backend/requirements.txt
# Django 核心框架
Django==4.2.7
djangorestframework==3.14.0

# 数据库
psycopg2-binary==2.9.7

# 认证和权限
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# 环境变量管理
python-decouple==3.8

# 时间处理
pytz==2023.3

# HTTP 客户端 (用于调用 fastGPT API)
requests==2.31.0

# 数据验证
cerberus==1.3.5

# 开发工具
django-extensions==3.2.3
ipython==8.15.0

# 测试
pytest==7.4.2
pytest-django==4.5.2

# 代码质量
flake8==6.1.0
black==23.9.1

# 生产环境
gunicorn==21.2.0
whitenoise==6.6.0